pyproject.toml
[project]
name = "the-invisible-criteria-of-peer-review"
version = "0.1.0"
description = "Codebase for 'The Invisible Criteria of Peer Review: Mapping the Divergence Between Institutional Guidelines and Reviewer Inquiries using Generative Topic Modeling'."
authors = [
    {name = "Ozer Tanrısever"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-core (>=1.3.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "art (>=6.5,<7.0)",
    "pymongo (>=4.12.1,<5.0.0)",
    "plotly (>=6.1.0,<7.0.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "huggingface-hub[hf-xet] (>=0.32.0,<0.33.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "nbformat (>=5.10.4,<6.0.0)",
    "umap-learn (>=0.5.7,<0.6.0)",
    "pyvis (>=0.3.2,<0.4.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "openreview-py (>=1.52.0,<2.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "langchain-google-genai (>=2.1.12,<3.0.0)",
    "kaleido (>=1.1.0,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "statsmodels (>=0.14.5,<0.15.0)",
    "gensim (>=4.4.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
nvitop = "^1.5.0"
jupyterlab = "^4.4.6"

